+++ rand_distr/src/poisson.rs
@@ -36,23 +36,15 @@ use rand::Rng;
 /// use rand_distr::{Poisson, Distribution};
 ///
 /// let poi = Poisson::new(2.0).unwrap();
-/// let v = poi.sample(&mut rand::thread_rng());
+/// let v: f64 = poi.sample(&mut rand::rng());
 /// println!("{} is from a Poisson(2) distribution", v);
 /// ```
 #[derive(Clone, Copy, Debug, PartialEq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-pub struct Poisson<F>
+pub struct Poisson<F>(Method<F>)
 where
     F: Float + FloatConst,
-    Standard: Distribution<F>,
-{
-    lambda: F,
-    // precalculated values
-    exp_lambda: F,
-    log_lambda: F,
-    sqrt_2lambda: F,
-    magic_val: F,
-}
+    Standard: Distribution<F>;
 
 /// Error type returned from [`Poisson::new`].
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -61,6 +53,8 @@ pub enum Error {
     ShapeTooSmall,
     /// `lambda = ∞` or `lambda = nan`
     NonFinite,
+    /// `lambda` is too large, see [Poisson::MAX_LAMBDA]
+    ShapeTooLarge,
 }
 
 impl fmt::Display for Error {
@@ -68,12 +62,59 @@ impl fmt::Display for Error {
         f.write_str(match self {
             Error::ShapeTooSmall => "lambda is not positive in Poisson distribution",
             Error::NonFinite => "lambda is infinite or nan in Poisson distribution",
+            Error::ShapeTooLarge => {
+                "lambda is too large in Poisson distribution, see Poisson::MAX_LAMBDA"
+            }
         })
     }
 }
 
 #[cfg(feature = "std")]
 impl std::error::Error for Error {}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub(crate) struct KnuthMethod<F> {
+    exp_lambda: F,
+}
+
+impl<F: Float> KnuthMethod<F> {
+    pub(crate) fn new(lambda: F) -> Self {
+        KnuthMethod {
+            exp_lambda: (-lambda).exp(),
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+struct RejectionMethod<F> {
+    lambda: F,
+    log_lambda: F,
+    sqrt_2lambda: F,
+    magic_val: F,
+}
+
+impl<F: Float> RejectionMethod<F> {
+    pub(crate) fn new(lambda: F) -> Self {
+        let log_lambda = lambda.ln();
+        let sqrt_2lambda = (F::from(2.0).unwrap() * lambda).sqrt();
+        let magic_val = lambda * log_lambda - crate::utils::log_gamma(F::one() + lambda);
+        RejectionMethod {
+            lambda,
+            log_lambda,
+            sqrt_2lambda,
+            magic_val,
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+enum Method<F> {
+    Knuth(KnuthMethod<F>),
+    Rejection(RejectionMethod<F>),
+}
 
 impl<F> Poisson<F>
 where
@@ -82,6 +123,8 @@ where
 {
     /// Construct a new `Poisson` with the given shape parameter
     /// `lambda`.
+    ///
+    /// The maximum allowed lambda is [MAX_LAMBDA](Self::MAX_LAMBDA).
     pub fn new(lambda: F) -> Result<Poisson<F>, Error> {
         if !lambda.is_finite() {
             return Err(Error::NonFinite);
@@ -89,14 +132,95 @@ where
         if !(lambda > F::zero()) {
             return Err(Error::ShapeTooSmall);
         }
-        let log_lambda = lambda.ln();
-        Ok(Poisson {
-            lambda,
-            exp_lambda: (-lambda).exp(),
-            log_lambda,
-            sqrt_2lambda: (F::from(2.0).unwrap() * lambda).sqrt(),
-            magic_val: lambda * log_lambda - crate::utils::log_gamma(F::one() + lambda),
-        })
+
+        // Use the Knuth method only for low expected values
+        let method = if lambda < F::from(12.0).unwrap() {
+            Method::Knuth(KnuthMethod::new(lambda))
+        } else {
+            if lambda > F::from(Self::MAX_LAMBDA).unwrap() {
+                return Err(Error::ShapeTooLarge);
+            }
+            Method::Rejection(RejectionMethod::new(lambda))
+        };
+
+        Ok(Poisson(method))
+    }
+
+    /// The maximum supported value of `lambda`
+    ///
+    /// This value was selected such that
+    /// `MAX_LAMBDA + 1e6 * sqrt(MAX_LAMBDA) < 2^64 - 1`,
+    /// thus ensuring that the probability of sampling a value larger than
+    /// `u64::MAX` is less than 1e-1000.
+    ///
+    /// Applying this limit also solves
+    /// [#1312](https://github.com/rust-random/rand/issues/1312).
+    pub const MAX_LAMBDA: f64 = 1.844e19;
+}
+
+impl<F> Distribution<F> for KnuthMethod<F>
+where
+    F: Float + FloatConst,
+    Standard: Distribution<F>,
+{
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> F {
+        let mut result = F::one();
+        let mut p = rng.random::<F>();
+        while p > self.exp_lambda {
+            p = p * rng.random::<F>();
+            result = result + F::one();
+        }
+        result - F::one()
+    }
+}
+
+impl<F> Distribution<F> for RejectionMethod<F>
+where
+    F: Float + FloatConst,
+    Standard: Distribution<F>,
+{
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> F {
+        // The algorithm from Numerical Recipes in C
+
+        // we use the Cauchy distribution as the comparison distribution
+        // f(x) ~ 1/(1+x^2)
+        let cauchy = Cauchy::new(F::zero(), F::one()).unwrap();
+        let mut result;
+
+        loop {
+            let mut comp_dev;
+
+            loop {
+                // draw from the Cauchy distribution
+                comp_dev = rng.sample(cauchy);
+                // shift the peak of the comparison distribution
+                result = self.sqrt_2lambda * comp_dev + self.lambda;
+                // repeat the drawing until we are in the range of possible values
+                if result >= F::zero() {
+                    break;
+                }
+            }
+            // now the result is a random variable greater than 0 with Cauchy distribution
+            // the result should be an integer value
+            result = result.floor();
+
+            // this is the ratio of the Poisson distribution to the comparison distribution
+            // the magic value scales the distribution function to a range of approximately 0-1
+            // since it is not exact, we multiply the ratio by 0.9 to avoid ratios greater than 1
+            // this doesn't change the resulting distribution, only increases the rate of failed drawings
+            let check = F::from(0.9).unwrap()
+                * (F::one() + comp_dev * comp_dev)
+                * (result * self.log_lambda
+                    - crate::utils::log_gamma(F::one() + result)
+                    - self.magic_val)
+                    .exp();
+
+            // check with uniform random value - if below the threshold, we are within the target distribution
+            if rng.random::<F>() <= check {
+                break;
+            }
+        }
+        result
     }
 }
 
@@ -107,60 +231,18 @@ where
 {
     #[inline]
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> F {
-        // using the algorithm from Numerical Recipes in C
-
-        // for low expected values use the Knuth method
-        if self.lambda < F::from(12.0).unwrap() {
-            let mut result = F::one();
-            let mut p = rng.random::<F>();
-            while p > self.exp_lambda {
-                p = p * rng.random::<F>();
-                result = result + F::one();
-            }
-            result - F::one()
+        match &self.0 {
+            Method::Knuth(method) => method.sample(rng),
+            Method::Rejection(method) => method.sample(rng),
         }
-        // high expected values - rejection method
-        else {
-            // we use the Cauchy distribution as the comparison distribution
-            // f(x) ~ 1/(1+x^2)
-            let cauchy = Cauchy::new(F::zero(), F::one()).unwrap();
-            let mut result;
+    }
+}
 
-            loop {
-                let mut comp_dev;
-
-                loop {
-                    // draw from the Cauchy distribution
-                    comp_dev = rng.sample(cauchy);
-                    // shift the peak of the comparison distribution
-                    result = self.sqrt_2lambda * comp_dev + self.lambda;
-                    // repeat the drawing until we are in the range of possible values
-                    if result >= F::zero() {
-                        break;
-                    }
-                }
-                // now the result is a random variable greater than 0 with Cauchy distribution
-                // the result should be an integer value
-                result = result.floor();
-
-                // this is the ratio of the Poisson distribution to the comparison distribution
-                // the magic value scales the distribution function to a range of approximately 0-1
-                // since it is not exact, we multiply the ratio by 0.9 to avoid ratios greater than 1
-                // this doesn't change the resulting distribution, only increases the rate of failed drawings
-                let check = F::from(0.9).unwrap()
-                    * (F::one() + comp_dev * comp_dev)
-                    * (result * self.log_lambda
-                        - crate::utils::log_gamma(F::one() + result)
-                        - self.magic_val)
-                        .exp();
-
-                // check with uniform random value - if below the threshold, we are within the target distribution
-                if rng.random::<F>() <= check {
-                    break;
-                }
-            }
-            result
-        }
+impl Distribution<u64> for Poisson<f64> {
+    #[inline]
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> u64 {
+        // `as` from float to int saturates
+        <Poisson<f64> as Distribution<f64>>::sample(self, rng) as u64
     }
 }